Started adding Controller Methods

+++ src/main/type/Nurse.java
@@ -1,11 +1,11 @@
 package main.type;
 
-import javafx.beans.property.IntegerProperty;
+import javafx.beans.property.LongProperty;
 import javafx.beans.property.StringProperty;
 
 public class Nurse extends Employee {
 
-    public Nurse(IntegerProperty id, StringProperty username, StringProperty password, StringProperty firstName, StringProperty lastName) {
+    public Nurse(LongProperty id, StringProperty username, StringProperty password, StringProperty firstName, StringProperty lastName) {
         super(id, username, password, firstName, lastName);
     }
 }
